Specify dopamine-rl version
For `dopamine-rl > 3.0.1`, it depends on `tensorflow >= 2.2.0`. That is why install this custom version of tensor2tensor will automatically upgrade the tensorflow version while SV2P still requires and old version, as described in https://github.com/suraj-nair-1/lorel.
In addition, `dopamine-rl > 3.0.1` requires `flax` and `jax`, but I believe this custom version of tensor2tensor doesn't depend on them.
After this PR, the installation of this package won't be tricky anymore.